UBUNTU: SAUCE: ubuntu_qrt*: use pre-rsync git repo for the test
Experiment shows that it will take about 1 ~ 20 minutes to clone the
qa-regression-testing repo for our *qrt* tests, it's all depends on
the connection speed to the launchpad. And it seems quite unstable.

Take ubuntu_qrt_kernel_aslr_collisions on Oracle VM.Standard2.1 for
example, the jenkins job took 35 minutes to complete, but this
lengthy git clone task occupied 1/3 of the time.

It just need ~30 seconds to copy it from our jenkins server to a SUT
with rsync. I think it's worthy to optimize this for the rest of qrt
related tests like what we did for ubuntu_qrt_kernel_hardening test
in commit 6441fc1e9b ("UBUNTU: SAUCE: ubuntu_qrt_kernel_hardening:
resolve lengthy git clones")

+++ ubuntu_qrt_apparmor/ubuntu_qrt_apparmor.py
@@ -76,8 +76,20 @@ class ubuntu_qrt_apparmor(test.test):
         self.install_required_pkgs()
 
         os.chdir(self.srcdir)
-        shutil.rmtree('qa-regression-testing', ignore_errors=True)
-        cmd = 'git clone --depth 1 https://git.launchpad.net/qa-regression-testing'
+        # Kernel QA Automation already copies the qa-regression-testing
+        # repo over to the SUT(system under test) via rsync+ssh.
+        # This resolves issues with extremely long git clones. Causing
+        # tests to fail.
+        # If qa-regression-testing exists in the SUT Homedir, just move
+        # it over to the autotest workarea. If not, then clone it
+        targetpath = os.path.expanduser("~") + "/qa-regression-testing"
+        if os.path.isdir(targetpath):
+            cmd = 'mv %s .' % targetpath
+        else:
+            # If the directory does not exist, then lets clone it as this test
+            # is probably being run by someone triaging a problem.
+            shutil.rmtree('qa-regression-testing', ignore_errors=True)
+            cmd = 'git clone --depth 1 https://git.launchpad.net/qa-regression-testing'
         self.results = utils.system_output(cmd, retain_output=True)
         # Print test suite HEAD SHA1 commit id for future reference
         os.chdir(os.path.join(self.srcdir, 'qa-regression-testing'))
